Rearrange functions and add ref corrections
- break related functions into categorized .py files and directories
- create fast linear least square poly fit routine
- write python-based reference pixel correction routine
